refactor: 자동로그인 처리 facade패턴으로 리팩터링
