+++ src/main/java/com/map/gaja/user/presentation/api/UserController.java
@@ -1,10 +1,7 @@
 package com.map.gaja.user.presentation.api;
 
-import com.map.gaja.client.apllication.ClientQueryService;
-import com.map.gaja.client.presentation.dto.response.ClientListResponse;
 import com.map.gaja.global.log.TimeCheckLog;
-import com.map.gaja.group.application.GroupService;
-import com.map.gaja.group.presentation.dto.response.GroupInfo;
+import com.map.gaja.user.application.AutoLoginProcessor;
 import com.map.gaja.user.application.UserService;
 import com.map.gaja.user.presentation.dto.request.LoginRequest;
 import com.map.gaja.user.presentation.api.specification.UserApiSpecification;
@@ -24,8 +21,7 @@ import javax.servlet.http.HttpSession;
 @RequiredArgsConstructor
 public class UserController implements UserApiSpecification {
     private final UserService userService;
-    private final ClientQueryService clientQueryService;
-    private final GroupService groupService;
+    private final AutoLoginProcessor autoLoginProcessor;
 
     @Override
     @PostMapping("/login")
@@ -51,23 +47,8 @@ public class UserController implements UserApiSpecification {
     @Override
     @GetMapping("/auto-login")
     public ResponseEntity<AutoLoginResponse> autoLogin(@AuthenticationPrincipal(expression = "name") String email) {
-        GroupInfo groupInfo = groupService.findGroup(email);
-
-        ClientListResponse clientListResponse;
-        if (isWholeGroup(groupInfo)) { //최근에 참조한 그룹이 전체일 경우
-            clientListResponse = clientQueryService.findAllClient(email, null);
-        } else { //최근에 참조한 그룹이 특정 그룹일 경우
-            clientListResponse = clientQueryService.findAllClientsInGroup(groupInfo.getGroupId(), null);
-        }
-
-        AutoLoginResponse response = new AutoLoginResponse(clientListResponse, groupInfo);
+        AutoLoginResponse response = autoLoginProcessor.process(email);
         return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
-    private boolean isWholeGroup(GroupInfo groupInfo) {
-        if (groupInfo == null) {
-            return true;
-        }
-        return false;
-    }
 }
